requirements.txt
Flask
python-dotenv
psycopg2-binary
PyMuPDF
pandas
